% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{runShinyPackageApp}
\alias{runShinyPackageApp}
\title{launch shiny app}
\usage{
runShinyPackageApp(launch_param = NULL, ...)
}
\arguments{
\item{launch_param}{param passed at launch time by user}

\item{...}{params are passed to shiny::runApp, e.g. port, launch.browser,
etc.}
}
\value{
This function normally does not return; interrupt R to stop the
  application (usually by pressing Ctrl+C or Esc).
}
\description{
Launch main shiny app in package with usual params of \code{shiny::runApp}
}
\examples{
\donttest{runShinyPackageApp(launch.browser = TRUE)}
}
