% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{shinypackage_server}
\alias{shinypackage_server}
\title{server function for shiny app}
\usage{
shinypackage_server(input, output)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}
}
\value{
a shiny server object
}
\description{
server function for shiny app
}
